Receive basic unit test
